Fix including ssh public keys
merge() cannot be used to override as this was causing create_resources
to blow up as the hash ended up looking like:

  {
    "key name": { "key": "some-rsa-key" },
    "ensure": "present",
    "user": "username"
  }

I have merged the defaults and override hashes so that the ensure and
username values will still be merged before creation. It is also
reasonable that a user of this module would want to override ensure.